removing old aws lambda serverless instrumentations
