--- tracer/src/Datadog.Trace/ClrProfiler/ServerlessInstrumentation/AWS/LambdaNoParamVoid.cs
@@ -1,46 +1,0 @@
-// <copyright file="LambdaNoParamVoid.cs" company="Datadog">
-// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
-// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
-// </copyright>
-
-using System;
-using System.ComponentModel;
-using Datadog.Trace.ClrProfiler.CallTarget;
-
-namespace Datadog.Trace.ClrProfiler.ServerlessInstrumentation.AWS
-{
-    /// <summary>
-    /// Lambda customer handler calltarget instrumentation
-    /// </summary>
-    [Browsable(false)]
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public class LambdaNoParamVoid
-    {
-        private static readonly ILambdaExtensionRequest RequestBuilder = new LambdaRequestBuilder();
-
-        /// <summary>
-        /// OnMethodBegin callback
-        /// </summary>
-        /// <typeparam name="TTarget">Type of the target</typeparam>
-        /// <param name="instance">Instance value, aka `this` of the instrumented method.</param>
-        /// <returns>Calltarget state value</returns>
-        internal static CallTargetState OnMethodBegin<TTarget>(TTarget instance)
-        {
-            Serverless.Debug("OnMethodBegin - no param");
-            return LambdaCommon.StartInvocationWithoutEvent(RequestBuilder);
-        }
-
-        /// <summary>
-        /// OnMethodEnd callback
-        /// </summary>
-        /// <typeparam name="TTarget">Type of the target</typeparam>
-        /// <param name="instance">Instance value, aka `this` of the instrumented method.</param>
-        /// <param name="exception">Exception instance in case the original code threw an exception.</param>
-        /// <param name="state">Calltarget state value</param>
-        internal static CallTargetReturn OnMethodEnd<TTarget>(TTarget instance, Exception exception, in CallTargetState state)
-        {
-            Serverless.Debug("OnMethodEnd - no param");
-            return LambdaCommon.EndInvocationVoid(exception, state.Scope, RequestBuilder);
-        }
-    }
-}
