--- tracer/src/Datadog.Trace/ClrProfiler/ServerlessInstrumentation/AWS/LambdaOneParamSync.cs
@@ -1,52 +1,0 @@
-// <copyright file="LambdaOneParamSync.cs" company="Datadog">
-// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
-// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
-// </copyright>
-
-using System;
-using System.ComponentModel;
-
-using Datadog.Trace.ClrProfiler.CallTarget;
-
-namespace Datadog.Trace.ClrProfiler.ServerlessInstrumentation.AWS
-{
-    /// <summary>
-    /// Lambda customer handler calltarget instrumentation
-    /// </summary>
-    [Browsable(false)]
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public class LambdaOneParamSync
-    {
-        private static readonly ILambdaExtensionRequest RequestBuilder = new LambdaRequestBuilder();
-
-        /// <summary>
-        /// OnMethodBegin callback
-        /// </summary>
-        /// <typeparam name="TTarget">Type of the target</typeparam>
-        /// <typeparam name="TArg">Type of the incomingEventOrContent</typeparam>
-        /// <param name="instance">Instance value, aka `this` of the instrumented method.</param>
-        /// <param name="incomingEventOrContext">IncomingEventOrContext value</param>
-        /// <returns>Calltarget state value</returns>
-        internal static CallTargetState OnMethodBegin<TTarget, TArg>(TTarget instance, TArg incomingEventOrContext)
-        {
-            Serverless.Debug("OnMethodBegin - one param");
-            return LambdaCommon.StartInvocationOneParameter(RequestBuilder, incomingEventOrContext);
-        }
-
-        /// <summary>
-        /// OnMethodEnd callback
-        /// </summary>
-        /// <typeparam name="TTarget">Type of the target</typeparam>
-        /// <typeparam name="TReturn">Type of the response, in an async scenario will be T of Task of T</typeparam>
-        /// <param name="instance">Instance value, aka `this` of the instrumented method.</param>
-        /// <param name="returnValue">HttpResponse message instance</param>
-        /// <param name="exception">Exception instance in case the original code threw an exception.</param>
-        /// <param name="state">Calltarget state value</param>
-        /// <returns>A response value</returns>
-        internal static CallTargetReturn<TReturn> OnMethodEnd<TTarget, TReturn>(TTarget instance, TReturn returnValue, Exception exception, in CallTargetState state)
-        {
-            Serverless.Debug("OnMethodEnd - one param");
-            return LambdaCommon.EndInvocationSync(returnValue, exception, state.Scope, RequestBuilder);
-        }
-    }
-}
