--- tracer/src/Datadog.Trace/ClrProfiler/ServerlessInstrumentation/AWS/LambdaTwoParamsAsync.cs
@@ -1,54 +1,0 @@
-// <copyright file="LambdaTwoParamsAsync.cs" company="Datadog">
-// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
-// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
-// </copyright>
-
-using System;
-using System.ComponentModel;
-using System.Threading.Tasks;
-using Datadog.Trace.ClrProfiler.CallTarget;
-
-namespace Datadog.Trace.ClrProfiler.ServerlessInstrumentation.AWS
-{
-    /// <summary>
-    /// Lambda customer handler calltarget instrumentation
-    /// </summary>
-    [Browsable(false)]
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public class LambdaTwoParamsAsync
-    {
-        private static readonly ILambdaExtensionRequest RequestBuilder = new LambdaRequestBuilder();
-
-        /// <summary>
-        /// OnMethodBegin callback
-        /// </summary>
-        /// <typeparam name="TTarget">Type of the target</typeparam>
-        /// <typeparam name="TArg1">Type of the incomingEvent</typeparam>
-        /// <typeparam name="TArg2">Type of the context</typeparam>
-        /// <param name="instance">Instance value, aka `this` of the instrumented method.</param>
-        /// <param name="incomingEvent">IncomingEvent value</param>
-        /// <param name="context">Context value.</param>
-        /// <returns>Calltarget state value</returns>
-        internal static CallTargetState OnMethodBegin<TTarget, TArg1, TArg2>(TTarget instance, TArg1 incomingEvent, TArg2 context)
-        {
-            Serverless.Debug("OnMethodBeginOK - two params");
-            return LambdaCommon.StartInvocationTwoParameters(RequestBuilder, incomingEvent, context);
-        }
-
-        /// <summary>
-        /// OnMethodEnd callback
-        /// </summary>
-        /// <typeparam name="TTarget">Type of the target</typeparam>
-        /// <typeparam name="TReturn">Type of the response, in an async scenario will be T of Task of T</typeparam>
-        /// <param name="instance">Instance value, aka `this` of the instrumented method.</param>
-        /// <param name="returnValue">HttpResponse message instance</param>
-        /// <param name="exception">Exception instance in case the original code threw an exception.</param>
-        /// <param name="state">Calltarget state value</param>
-        /// <returns>A response value</returns>
-        internal static Task<TReturn> OnAsyncMethodEnd<TTarget, TReturn>(TTarget instance, TReturn returnValue, Exception exception, in CallTargetState state)
-        {
-            Serverless.Debug("OnMethodEnd - two params");
-            return LambdaCommon.EndInvocationAsync(returnValue, exception, state.Scope, RequestBuilder);
-        }
-    }
-}
